refactor: 统一分割线为 divide-light 快捷方式，将 DisplaySettings 内联并删除旧 Modal 包装组件
