--- src/renderer/components/SettingsModal/contents/DisplayModalContent.tsx
@@ -1,69 +1,0 @@
-/**
- * @license
- * Copyright 2025 Sudowork (sudowork.ai)
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import React from 'react';
-import { useTranslation } from 'react-i18next';
-import FontSizeControl from '@/renderer/components/FontSizeControl';
-import { ThemeSwitcher } from '@/renderer/components/ThemeSwitcher';
-import AionScrollArea from '@/renderer/components/base/AionScrollArea';
-
-/**
- * 偏好设置行组件 / Preference row component
- * 用于显示标签和对应的控件，统一的水平布局 / Used for displaying labels and corresponding controls in a unified horizontal layout
- */
-const PreferenceRow: React.FC<{
-  /** 标签文本 / Label text */
-  label: string;
-  /** 控件元素 / Control element */
-  children: React.ReactNode;
-}> = ({ label, children }) => (
-  <div className='flex flex-col items-stretch gap-10px py-12px md:flex-row md:items-center md:justify-between md:gap-24px'>
-    <div className='text-14px text-t-primary leading-22px'>{label}</div>
-    <div className='w-full flex md:flex-1 md:justify-end'>{children}</div>
-  </div>
-);
-
-/**
- * 显示设置内容组件 / Display settings content component
- *
- * 提供显示相关的配置选项，包括主题和缩放比例
- * Provides display-related configuration options including theme and zoom scale
- *
- * @features
- * - 主题切换：亮色/暗色/跟随系统 / Theme: light/dark/system
- * - 缩放比例控制 / Zoom scale control
- */
-const DisplayModalContent: React.FC = () => {
-  const { t } = useTranslation();
-
-  // 显示设置项配置 / Display items configuration
-  const displayItems = [
-    { key: 'theme', label: t('settings.theme'), component: <ThemeSwitcher /> },
-    { key: 'fontSize', label: t('settings.fontSize'), component: <FontSizeControl /> },
-  ];
-
-  return (
-    <div className='flex flex-col h-full w-full'>
-      {/* 内容区域 / Content Area */}
-      <AionScrollArea className='flex-1 min-h-0 pb-16px' disableOverflow>
-        <div className='space-y-16px'>
-          {/* 显示设置 / Display Settings */}
-          <div className='px-16px md:px-24px lg:px-28px py-14px md:py-16px bg-2 rd-16px space-y-10px md:space-y-12px'>
-            <div className='w-full flex flex-col divide-y divide-[var(--border-light)]'>
-              {displayItems.map((item) => (
-                <PreferenceRow key={item.key} label={item.label}>
-                  {item.component}
-                </PreferenceRow>
-              ))}
-            </div>
-          </div>
-        </div>
-      </AionScrollArea>
-    </div>
-  );
-};
-
-export default DisplayModalContent;
