+++ src/renderer/components/SettingsModal/contents/SystemModalContent.tsx
@@ -445,7 +445,7 @@ const SystemModalContent: React.FC = () => {
         <div className='space-y-16px'>
           {/* 偏好设置与高级设置合并展示 / Combined preferences and advanced settings */}
           <div className='px-[12px] md:px-[32px] py-16px bg-2 rd-16px space-y-12px'>
-            <div className='w-full flex flex-col divide-y divide-[var(--border-light)]'>
+            <div className='w-full flex flex-col divide-y divide-light'>
               {preferenceItems.map((item) => (
                 <PreferenceRow key={item.key} label={item.label} hint={item.hint}>
                   {item.component}
