+++ src/renderer/pages/settings/DisplaySettings.tsx
@@ -5,13 +5,49 @@
  */
 
 import React from 'react';
-import DisplayModalContent from '@/renderer/components/SettingsModal/contents/DisplayModalContent';
+import { useTranslation } from 'react-i18next';
+import FontSizeControl from '@/renderer/components/FontSizeControl';
+import { ThemeSwitcher } from '@/renderer/components/ThemeSwitcher';
+import AionScrollArea from '@/renderer/components/base/AionScrollArea';
 import SettingsPageWrapper from './components/SettingsPageWrapper';
 
+const PreferenceRow: React.FC<{
+  label: string;
+  children: React.ReactNode;
+}> = ({ label, children }) => (
+  <div className='flex flex-col items-stretch gap-2.5 py-3 md:flex-row md:items-center md:justify-between md:gap-6'>
+    <div className='text-14px text-t-primary leading-22px'>{label}</div>
+    <div className='w-full flex md:flex-1 md:justify-end'>{children}</div>
+  </div>
+);
+
 const DisplaySettings: React.FC = () => {
+  const { t } = useTranslation();
+
+  // 显示设置项配置 / Display items configuration
+  const displayItems = [
+    { key: 'theme', label: t('settings.theme'), component: <ThemeSwitcher /> },
+    { key: 'fontSize', label: t('settings.fontSize'), component: <FontSizeControl /> },
+  ];
+
   return (
     <SettingsPageWrapper>
-      <DisplayModalContent />
+      <div className='flex flex-col h-full w-full'>
+        {/* 内容区域 / Content Area */}
+        <AionScrollArea className='flex-1 min-h-0 pb-4' disableOverflow>
+          <div className='space-y-4'>
+            <div className='px-4 md:px-6 lg:px-7 py-3.5 md:py-4 bg-2 rd-16px space-y-2.5 md:space-y-3'>
+              <div className='w-full flex flex-col divide-y divide-light'>
+                {displayItems.map((item) => (
+                  <PreferenceRow key={item.key} label={item.label}>
+                    {item.component}
+                  </PreferenceRow>
+                ))}
+              </div>
+            </div>
+          </div>
+        </AionScrollArea>
+      </div>
     </SettingsPageWrapper>
   );
 };
