+++ uno.config.ts
@@ -33,10 +33,10 @@ const semanticColors = {
 // 用途：仅用于背景、容器等布局元素（bg-*）。
 // Usage: backgrounds / containers only (bg-*).
 // ⚠️ 这些是背景色，请勿当边框用 —— 边框统一走语义 token：
-//    border-default / border-subtle 快捷方式，或 border-[var(--border-default)] / border-[var(--border-light)]。
+//    border-light 快捷方式（浅边框），或 border-[var(--border-default)] / border-[var(--border-light)]。
 //    UnoCSS 颜色命名空间共享，border-1 / border-base 等技术上能解析，但指向的是「背景色」而非边框色，属误用。
 // ⚠️ These are background colors — do NOT use as borders. For borders use the semantic
-//    shortcuts border-default / border-subtle, or border-[var(--border-default|light)].
+//    shortcut border-light (lighter border), or border-[var(--border-default|light)].
 // 📝 text-1 到 text-4 通过自定义规则支持，指向 Arco 的 --color-text-*
 // text-1 to text-4 are supported via custom rules, pointing to Arco's --color-text-*
 const backgroundColors = {
@@ -138,14 +138,8 @@ export default defineConfig({
   // 基础配置
   shortcuts: {
     'f-center': 'flex items-center justify-center',
-    // 语义化边框颜色 / Semantic border colors
-    // 标准边框无需快捷方式：preflight 已把 --border-default 设为全局默认边框色，
-    // 直接写宽度类即可（border / border-b = 1px solid 默认色边框）。
-    // 此处仅为「浅边框」提供命名，避免写裸 border-[var(--border-light)]。
-    // Standard border needs no shortcut — preflight defaults border-color to --border-default,
-    // so a width class alone (border / border-b) already gives a 1px solid default border.
-    // Only the lighter variant gets a name here. 仅产出 border-color，搭配宽度类使用。
-    'border-subtle': 'border-[var(--border-light)]', // 浅边框 / lighter divider（搭配 border / border-b）
+    'border-light': 'border-[var(--border-light)]', // 浅边框 / lighter border（搭配 border / border-b）
+    'divide-light': 'divide-light', // 浅分割线 / lighter divider color（搭配 divide-y / divide-x）
     'scrollbar-hide': 'scrollbar-width-none [&::-webkit-scrollbar]:hidden',
     // 技能/数字助手商店卡片：悬浮白卡 + hover 抬升 / elevated white store card
     'library-card': 'group bg-fill-0 rd-12px p-12px flex items-start gap-12px relative overflow-hidden shadow-[0_1px_3px_rgba(0,0,0,0.08)] transition-all duration-200 hover:-translate-y-px hover:shadow-[0_4px_12px_rgba(0,0,0,0.1)]',
